Changed the collision system to use ints instead of floats and got rid of the player collision system

+++ src/Components/BoxColliderComponent.h
@@ -5,7 +5,7 @@
 struct BoxColliderComponent
 {
     Vector2 position;
-    float w;
-    float h;
+    int w;
+    int h;
     bool isBottomCollision = false;
 };
